modified:   app/Views/product/hotel/hotel-details.php

+++ app/Views/product/hotel/hotel-details.php
@@ -1137,10 +1137,10 @@
                                 </td>
 								
 								<?php
-								       $basic_won  =  $room['goods_price1'] * $room['baht_thai'];
+								       $basic_won  =  (int)($room['goods_price1'] * $room['baht_thai']);
 								       $basic_bath =  $room['goods_price1'];
 									   
-								       $price_won  = ($room['goods_price2'] + $room['goods_price3']) * $room['baht_thai'];
+								       $price_won  =  (int)(($room['goods_price2'] + $room['goods_price3']) * $room['baht_thai']);
 								       $price_bath =  $room['goods_price2'] + $room['goods_price3'];
 									   
 								?>
